Fixing renaming issues with footer and header
