+++ src/app/contact/page.tsx
@@ -1,10 +1,12 @@
-import Header from "@/components/Header";
-import Footer from "@/components/Footer";
+import Header from "@/components/SesaHeader";
+import Footer from "@/components/SesaFooter";
+import ContactForm from "@/components/ContactForm";
 
 export default function ContactPage() {
   return (
     <div>
       <Header />
+      <ContactForm />
       <Footer />
     </div>
   );
