+++ src/app/page.tsx
@@ -1,7 +1,7 @@
 "use client";
 
-import Header from "@/components/Header";
-import Footer from "@/components/Footer";
+import Header from "@/components/SesaHeader";
+import Footer from "@/components/SesaFooter";
 import SplashHeader from "@/components/SplashHeader";
 
 import HomeBackground from "../../public/homeBackgroundGreen.png";
