+++ src/app/sponsors/page.tsx
@@ -1,5 +1,5 @@
-import Header from "@/components/Header";
-import Footer from "@/components/Footer";
+import Header from "@/components/SesaHeader";
+import Footer from "@/components/SesaFooter";
 
 export default function SponsorsPage() {
   return (
